+++ Ticket_Center/src/main/java/bg/tu_varna/sit/group25/ticketcenter/Application/presentation/models/ShowListViewModel.java
@@ -17,10 +17,10 @@ public class ShowListViewModel
     public int tickets;
     public long status;
     public String s_type;
-    public long organizer;
-    public long distributor;
+    public int organizer;
+    public int distributor;
 
-    public ShowListViewModel(int s_id, String title, GregorianCalendar date, String location, int max, int tickets, int status, String s_type, int organizer, long distributor) {
+    public ShowListViewModel(int s_id, String title, GregorianCalendar date, String location, int max, int tickets, int status, String s_type, int organizer, int distributor) {
         this.s_id = s_id;
         this.title = title;
         this.date = date;
